acm-requirements.txt
psycopg2-binary~=2.7.4
python-dotenv~=0.20.0
requests~=2.27.1
beautifulsoup4~=4.10.0
dataclasses~=0.8